+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Emit all types into the generated `GraphQLSchema` explicity.
 - Also allow `Null.t` to be used as is, in addition to `Nullable.t`.
+- Don't validate the number of fields in types, let the GraphQL server do that instead.
 
 ## 0.10.0